requirements.txt
anthropic>=0.18.0
openai>=1.0.0
keyboard>=0.13.5
pyperclip>=1.8.2
python-dotenv>=1.0.0
pyyaml>=6.0
pyautogui>=0.9.54
requests>=2.31.0
